Create File 6_Temperature Check Survey
File 6_Temperature Check Survey 

Instructions: Approximately 3-6 months after a community death, it’s good to check in to see how people are doing and assess if any additional response is needed or if the postvention plan can be improved. Reaching out directly to community members most impacted first is highly recommended to determine if they have continued or additional support.  Following this, sending a departmental email might be reasonable and can be used to achieve the following components:

1. Share resources again 
2. Gather information for improvement of the suicide postvention plan 
3. Determine if there isa. community need for continued response and support
3. Communicate how the department may respond to the newly collected information* 

* It is best practice to have clear action steps that you are able/willing to take after collecting information from the community.  Not having a response after collecting sensitive information or reactions may be seen as a choice and can diminish morale.