Create File_7 Guidelines for long-term support for close-proximity colleagues
File_7 Guidelines for long-term support for close-proximity colleagues

Instructions: It is important to consider and be prepared to take action to support close-proximity colleagues. In academic and research settings, there are potential long-term issues related to graduation dates, anniversaries, publications that the person contributed to, persistence in University-controlled systems (i.e. remains listed as a graduate student thesis advisee years after death), transfer of projects to others that may result in additional emotional distress for various community members.  These long-term considerations may be helpful to consider and act on both for faculty leaders of labs or groups affected and the departmental leaders.  

Sections: 
A.	Comfort 
B.	Restore
C.	Lead
D.	Honor
E.	Sustain